Working libft

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: brpereir <brpereir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 23:01:33 by brpereir          #+#    #+#             */
-/*   Updated: 2023/04/23 18:53:59 by brpereir         ###   ########.fr       */
+/*   Updated: 2023/04/24 19:32:23 by brpereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 	char	*str;
 
 	if (!s1 || !s2)
