+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: brpereir <brpereir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 19:56:06 by brpereir          #+#    #+#             */
-/*   Updated: 2023/04/22 15:02:18 by brpereir         ###   ########.fr       */
+/*   Updated: 2023/04/24 19:36:39 by brpereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char const *str)
+size_t	ft_strlen(char const *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
